+++ packages/core/src/__tests__/Matchers/query-string.test.js
@@ -8,8 +8,8 @@ describe('query string matching', () => {
 			response: 200,
 		});
 
-		expect(route.matcher('http://a.com')).toBe(false);
-		expect(route.matcher('http://a.com?a=b&c=d')).toBe(true);
+		expect(route.matcher({ url: 'http://a.com' })).toBe(false);
+		expect(route.matcher({ url: 'http://a.com?a=b&c=d' })).toBe(true);
 	});
 
 	it('match a query string against a URL object', () => {
@@ -20,7 +20,7 @@ describe('query string matching', () => {
 		const url = new URL('http://a.com/path');
 		url.searchParams.append('a', 'b');
 		url.searchParams.append('c', 'd');
-		expect(route.matcher(url)).toBe(true);
+		expect(route.matcher({ url })).toBe(true);
 	});
 
 	it('match a query string against a relative path', () => {
@@ -29,7 +29,7 @@ describe('query string matching', () => {
 			response: 200,
 		});
 		const url = '/path?a=b';
-		expect(route.matcher(url)).toBe(true);
+		expect(route.matcher({ url })).toBe(true);
 	});
 
 	it('match multiple query strings', () => {
@@ -38,10 +38,10 @@ describe('query string matching', () => {
 			response: 200,
 		});
 
-		expect(route.matcher('http://a.com')).toBe(false);
-		expect(route.matcher('http://a.com?a=b')).toBe(false);
-		expect(route.matcher('http://a.com?a=b&c=d')).toBe(true);
-		expect(route.matcher('http://a.com?c=d&a=b')).toBe(true);
+		expect(route.matcher({ url: 'http://a.com' })).toBe(false);
+		expect(route.matcher({ url: 'http://a.com?a=b' })).toBe(false);
+		expect(route.matcher({ url: 'http://a.com?a=b&c=d' })).toBe(true);
+		expect(route.matcher({ url: 'http://a.com?c=d&a=b' })).toBe(true);
 	});
 
 	it('ignore irrelevant query strings', () => {
@@ -50,7 +50,7 @@ describe('query string matching', () => {
 			response: 200,
 		});
 
-		expect(route.matcher('http://a.com?a=b&c=d&e=f')).toBe(true);
+		expect(route.matcher({ url: 'http://a.com?a=b&c=d&e=f' })).toBe(true);
 	});
 	it('match an empty query string', () => {
 		const route = new Route({
@@ -58,8 +58,8 @@ describe('query string matching', () => {
 			response: 200,
 		});
 
-		expect(route.matcher('http://a.com')).toBe(false);
-		expect(route.matcher('http://a.com?a=')).toBe(true);
+		expect(route.matcher({ url: 'http://a.com' })).toBe(false);
+		expect(route.matcher({ url: 'http://a.com?a=' })).toBe(true);
 	});
 
 	describe('value coercion', () => {
@@ -70,8 +70,8 @@ describe('query string matching', () => {
 				},
 				response: 200,
 			});
-			expect(route.matcher('http://a.com')).toBe(false);
-			expect(route.matcher('http://a.com?a=1')).toBe(true);
+			expect(route.matcher({ url: 'http://a.com' })).toBe(false);
+			expect(route.matcher({ url: 'http://a.com?a=1' })).toBe(true);
 		});
 
 		it('coerce floats to strings and match', () => {
@@ -81,8 +81,8 @@ describe('query string matching', () => {
 				},
 				response: 200,
 			});
-			expect(route.matcher('http://a.com')).toBe(false);
-			expect(route.matcher('http://a.com?a=1.2')).toBe(true);
+			expect(route.matcher({ url: 'http://a.com' })).toBe(false);
+			expect(route.matcher({ url: 'http://a.com?a=1.2' })).toBe(true);
 		});
 
 		it('coerce booleans to strings and match', () => {
@@ -99,10 +99,10 @@ describe('query string matching', () => {
 				response: 200,
 			});
 
-			expect(trueRoute.matcher('http://a.com')).toBe(false);
-			expect(falseRoute.matcher('http://a.com')).toBe(false);
-			expect(trueRoute.matcher('http://a.com?a=true')).toBe(true);
-			expect(falseRoute.matcher('http://a.com?b=false')).toBe(true);
+			expect(trueRoute.matcher({ url: 'http://a.com' })).toBe(false);
+			expect(falseRoute.matcher({ url: 'http://a.com' })).toBe(false);
+			expect(trueRoute.matcher({ url: 'http://a.com?a=true' })).toBe(true);
+			expect(falseRoute.matcher({ url: 'http://a.com?b=false' })).toBe(true);
 		});
 
 		it('coerce undefined to an empty string and match', () => {
@@ -112,8 +112,8 @@ describe('query string matching', () => {
 				},
 				response: 200,
 			});
-			expect(route.matcher('http://a.com')).toBe(false);
-			expect(route.matcher('http://a.com?a=')).toBe(true);
+			expect(route.matcher({ url: 'http://a.com' })).toBe(false);
+			expect(route.matcher({ url: 'http://a.com?a=' })).toBe(true);
 		});
 
 		it('coerce null to an empty string and match', () => {
@@ -123,8 +123,8 @@ describe('query string matching', () => {
 				},
 				response: 200,
 			});
-			expect(route.matcher('http://a.com')).toBe(false);
-			expect(route.matcher('http://a.com?a=')).toBe(true);
+			expect(route.matcher({ url: 'http://a.com' })).toBe(false);
+			expect(route.matcher({ url: 'http://a.com?a=' })).toBe(true);
 		});
 
 		it('coerce an object to an empty string and match', () => {
@@ -134,8 +134,8 @@ describe('query string matching', () => {
 				},
 				response: 200,
 			});
-			expect(route.matcher('http://a.com')).toBe(false);
-			expect(route.matcher('http://a.com?a=')).toBe(true);
+			expect(route.matcher({ url: 'http://a.com' })).toBe(false);
+			expect(route.matcher({ url: 'http://a.com?a=' })).toBe(true);
 		});
 
 		it('can match a query string with different value types', () => {
@@ -150,9 +150,11 @@ describe('query string matching', () => {
 				},
 			});
 
-			expect(route.matcher('http://a.com')).toBe(false);
+			expect(route.matcher({ url: 'http://a.com' })).toBe(false);
 			expect(
-				route.matcher('http://a.com?t=true&f=false&u=&num=1&arr=a&arr='),
+				route.matcher({
+					url: 'http://a.com?t=true&f=false&u=&num=1&arr=a&arr=',
+				}),
 			).toBe(true);
 		});
 	});
@@ -162,26 +164,26 @@ describe('query string matching', () => {
 		it('match repeated query strings', () => {
 			const route = new Route({ query: { a: ['b', 'c'] }, response: 200 });
 
-			expect(route.matcher('http://a.com')).toBe(false);
-			expect(route.matcher('http://a.com?a=b')).toBe(false);
-			expect(route.matcher('http://a.com?a=b&a=c')).toBe(true);
-			expect(route.matcher('http://a.com?a=b&a=c&a=d')).toBe(false);
+			expect(route.matcher({ url: 'http://a.com' })).toBe(false);
+			expect(route.matcher({ url: 'http://a.com?a=b' })).toBe(false);
+			expect(route.matcher({ url: 'http://a.com?a=b&a=c' })).toBe(true);
+			expect(route.matcher({ url: 'http://a.com?a=b&a=c&a=d' })).toBe(false);
 		});
 
 		it('match repeated query strings in any order', () => {
 			const route = new Route({ query: { a: ['b', 'c'] }, response: 200 });
 
-			expect(route.matcher('http://a.com')).toBe(false);
-			expect(route.matcher('http://a.com?a=b&a=c')).toBe(true);
-			expect(route.matcher('http://a.com?a=c&a=b')).toBe(true);
+			expect(route.matcher({ url: 'http://a.com' })).toBe(false);
+			expect(route.matcher({ url: 'http://a.com?a=b&a=c' })).toBe(true);
+			expect(route.matcher({ url: 'http://a.com?a=c&a=b' })).toBe(true);
 		});
 
 		it('match a query string array of length 1', () => {
 			const route = new Route({ query: { a: ['b'] }, response: 200 });
 
-			expect(route.matcher('http://a.com')).toBe(false);
-			expect(route.matcher('http://a.com?a=b')).toBe(true);
-			expect(route.matcher('http://a.com?a=b&a=c')).toBe(false);
+			expect(route.matcher({ url: 'http://a.com' })).toBe(false);
+			expect(route.matcher({ url: 'http://a.com?a=b' })).toBe(true);
+			expect(route.matcher({ url: 'http://a.com?a=b&a=c' })).toBe(false);
 		});
 
 		it('match a repeated query string with an empty value', () => {
@@ -190,9 +192,9 @@ describe('query string matching', () => {
 				response: 200,
 			});
 
-			expect(route.matcher('http://a.com')).toBe(false);
-			expect(route.matcher('http://a.com?a=b')).toBe(false);
-			expect(route.matcher('http://a.com?a=b&a=')).toBe(true);
+			expect(route.matcher({ url: 'http://a.com' })).toBe(false);
+			expect(route.matcher({ url: 'http://a.com?a=b' })).toBe(false);
+			expect(route.matcher({ url: 'http://a.com?a=b&a=' })).toBe(true);
 		});
 	});
 
@@ -205,10 +207,10 @@ describe('query string matching', () => {
 				query: { a: 'b' },
 			});
 
-			expect(route.matcher('http://a.com?c=d')).toBe(false);
-			expect(route.matcher('http://a.com?a=b')).toBe(false);
-			expect(route.matcher('http://a.com?c=d&a=b')).toBe(true);
-			expect(route.matcher('http://a.com?a=b&c=d')).toBe(true);
+			expect(route.matcher({ url: 'http://a.com?c=d' })).toBe(false);
+			expect(route.matcher({ url: 'http://a.com?a=b' })).toBe(false);
+			expect(route.matcher({ url: 'http://a.com?c=d&a=b' })).toBe(true);
+			expect(route.matcher({ url: 'http://a.com?a=b&c=d' })).toBe(true);
 		});
 
 		it('can be used alongside function matchers', () => {
@@ -218,8 +220,8 @@ describe('query string matching', () => {
 				query: { a: 'b' },
 			});
 
-			expect(route.matcher('http://a.com')).toBe(false);
-			expect(route.matcher('http://a.com?a=b')).toBe(true);
+			expect(route.matcher({ url: 'http://a.com' })).toBe(false);
+			expect(route.matcher({ url: 'http://a.com?a=b' })).toBe(true);
 		});
 	});
 });